easystore commit

+++ src/main/java/cn/usually/modules/controllers/platform/easystore/EasyProductCategoryController.java
@@ -7,11 +7,11 @@ import cn.usually.common.page.DataTableColumn;
 import cn.usually.common.page.DataTableOrder;
 import cn.usually.modules.models.platform.easystore.Easy_product_category;
 import cn.usually.modules.services.platform.easystore.EasyProductCategoryService;
+import cn.usually.modules.services.platform.easystore.EasyProductService;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.nutz.dao.Cnd;
 import org.nutz.ioc.loader.annotation.Inject;
 import org.nutz.ioc.loader.annotation.IocBean;
-import org.nutz.lang.Strings;
 import org.nutz.log.Log;
 import org.nutz.log.Logs;
 import org.nutz.mvc.annotation.*;
@@ -26,6 +26,9 @@ public class EasyProductCategoryController {
 	private static final Log log = Logs.get();
 	@Inject
 	private EasyProductCategoryService easyProductCategoryService;
+
+	@Inject
+	private EasyProductService easyProductService;
 
 	@At("")
 	@Ok("beetl:/platform/easystore/product/category/index.html")
@@ -89,11 +92,17 @@ public class EasyProductCategoryController {
     public Object delete(long id, @Param("ids") Long[] ids ,HttpServletRequest req) {
 		try {
 			if(ids!=null&&ids.length>0){
-				easyProductCategoryService.delete(ids);
-    			req.setAttribute("id", org.apache.shiro.util.StringUtils.toString(ids));
+				if(checkCanDelete(ids)) {
+					easyProductCategoryService.delete(ids);
+					req.setAttribute("id", org.apache.shiro.util.StringUtils.toString(ids));
+				}else
+					return Result.error("无法删除了，待删除产品类型中存在关联产品");
 			}else{
-				easyProductCategoryService.delete(id);
-    			req.setAttribute("id", id);
+				if(checkCanDelete(id)) {
+					easyProductCategoryService.delete(id);
+					req.setAttribute("id", id);
+				}else
+					return Result.error("无法删除了，该产品类型下存在关联产品");
 			}
 			return Result.success("system.success");
 		} catch (Exception e) {
@@ -101,8 +110,32 @@ public class EasyProductCategoryController {
 		}
     }
 
+	/**
+	 * 是否可以刪除多个产品类型
+	 * @param ids
+	 * @return
+	 */
+	private boolean checkCanDelete(Long[] ids) {
+		for (long id : ids) {
+			if(! checkCanDelete(id)) {
+				return false;
+			}
+		}
+		return true;
+	}
 
-    @At("/detail/?")
+	/**
+	 * 是否可以刪除单个产品类型:先查找待删除的产品类型下商品数量,>0则返回true;否则返回false
+	 * @param id
+	 * @return
+	 */
+	private boolean checkCanDelete(long id) {
+		long count = easyProductService.count(Cnd.where("category_id","=",id));
+		return count > 0 ? false : true;
+	}
+
+
+	@At("/detail/?")
     @Ok("beetl:/platform/easystore/product/category/detail.html")
     @RequiresAuthentication
 	public Object detail(long id) {
